Cleaned up package structure, updated changelog, bumped version
