+++ lib/src/requests/responses/rest_response.dart
@@ -1,4 +1,4 @@
-import '../../restrr.dart';
+import '../../../restrr.dart';
 
 /// Represents a response from a REST request.
 /// This can either hold data, [T], or an [ErrorResponse].
