+++ pubspec.yaml
@@ -1,6 +1,6 @@
 name: restrr
 description: Dart package which allows to communicate with the financrr REST API.
-version: 0.1.1
+version: 0.2.0
 repository: https://github.com/financrr/restrr
 
 environment:
